+++ tower/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **builder**: Add `ServiceBuilder::map_future` for transforming the futures produced
   by a service.
+- **builder**: Add `ServiceBuilder::service_fn` for applying `Layer` to an
+  async function using `util::service_fn`.
 
 # 0.4.5 (February 10, 2021)